pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "librar"
version = "0.1.0"
description = "Local-first book ingestion and search foundations"
requires-python = ">=3.11"
dependencies = [
    "pymupdf==1.26.7",
    "EbookLib==0.20",
    "lxml==6.0.2",
    "beautifulsoup4==4.14.3",
    "charset-normalizer==3.4.4",
    "faiss-cpu==1.13.0",
    "numpy==1.26.4",
    "openai==1.47.1",
    "python-dotenv>=1.0.1",
    "python-telegram-bot==22.6",
    "pymorphy2==0.9.1",
    "pymorphy2-dicts-ru==2.4.417127.4579844",
    "razdel==0.5.0",
    "watchdog>=4.0.0",
    "pytesseract>=0.3.13",
    "Pillow>=10.4.0",
    "lingua-language-detector>=2.0.2",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
classify-books = "librar.cli.classify_books:main"
build-timeline = "librar.cli.build_timeline:main"
